+++ packages/core/src/model/GraphModel.ts
@@ -17,6 +17,7 @@ import { snapToGrid, getGridOffset } from '../util/geometry';
 import { isPointInArea } from '../util/graph';
 import { getClosestPointOfPolyline } from '../util/edge';
 import { formatData } from '../util/compatible';
+import { getNodeAnchorPosition } from '../util/node';
 
 type BaseNodeModelId = string; // 节点ID
 type BaseEdgeModelId = string; // 连线ID
@@ -87,7 +88,10 @@ class GraphModel {
   }
   // fixme: 用户点击触发两次sortElements
   @computed get sortElements() {
-    const elements = [...this.edges, ...this.nodes];
+    const elements = [];
+    // IE BUG: mobx observer对象使用解构会导致IE11出现问题
+    this.nodes.forEach(node => elements.push(node));
+    this.edges.forEach(edge => elements.push(edge));
     // 只显示可见区域的节点和连线以及和这个可以区域节点的节点
     const showElements = [];
     let topElementIdx = -1;
@@ -125,7 +129,10 @@ class GraphModel {
    */
   getAreaElement(leftTopPoint, rightBottomPoint) {
     const selectElements = [];
-    const elements = [...this.edges, ...this.nodes];
+    const elements = [];
+    // IE BUG: mobx observer对象使用解构会导致IE11出现问题
+    this.nodes.forEach(node => elements.push(node));
+    this.edges.forEach(edge => elements.push(edge));
     for (let i = 0; i < elements.length; i++) {
       const currentItem = elements[i];
       if (this.isElementInArea(currentItem, leftTopPoint, rightBottomPoint)) {
@@ -229,10 +236,7 @@ class GraphModel {
   }
 
   getEdgeModel(edgeId: string) {
-    const edge = this.edgesMap[edgeId];
-    if (edge) {
-      return edge.model;
-    }
+    return this.edgesMap[edgeId]?.model;
   }
 
   getElement(id: string): IBaseModel | undefined {
@@ -376,7 +380,12 @@ class GraphModel {
   @action
   moveNode(nodeId: BaseNodeModelId, deltaX: number, deltaY: number) {
     // 1) 移动节点
-    const nodeModel = this.nodesMap[nodeId].model;
+    const node = this.nodesMap[nodeId];
+    if (!node) {
+      console.warn(`不存在id为${nodeId}的节点`);
+      return;
+    }
+    const nodeModel = node.model;
     nodeModel.move(deltaX, deltaY);
     // 2) 移动连线
     this.moveEdge(nodeId, deltaX, deltaY);
@@ -460,6 +469,11 @@ class GraphModel {
   @action
   removeEdgeById(id) {
     const idx = this.edgesMap[id].index;
+    const edge = this.edgesMap[id];
+    if (!edge) {
+      console.warn(`不存在id为${id}的边`);
+      return;
+    }
     const edgeData = this.edgesMap[id].model.getData();
     this.edges.splice(idx, 1);
     this.eventCenter.emit(EventType.EDGE_DELETE, { data: edgeData });
@@ -595,6 +609,46 @@ class GraphModel {
   changeEdgeType(type: string): void {
     this.edgeType = type;
   }
+  @action
+  changeNodeType(id, type: string): void {
+    console.log(id, type);
+    const nodeModel = this.getNodeModel(id);
+    if (!nodeModel) {
+      console.warn(`找不到id为${id}的节点`);
+      return;
+    }
+    const data = nodeModel.getData();
+    data.type = type;
+    const Model = this.getModel(type);
+    if (!Model) {
+      throw new Error(`找不到${type}对应的节点，请确认是否已注册此类型节点。`);
+    }
+    const newNodeModel = new Model(data, this);
+    this.nodes.splice(this.nodesMap[id].index, 1, newNodeModel);
+    // 微调连线
+    const edgeModels = this.getNodeEdges(id);
+    edgeModels.forEach(edge => {
+      if (edge.sourceNodeId === id) {
+        const point = getNodeAnchorPosition(
+          newNodeModel,
+          edge.startPoint,
+          newNodeModel.width,
+          newNodeModel.height,
+        );
+        edge.updateStartPoint(point);
+      }
+      if (edge.targetNodeId === id) {
+        const point = getNodeAnchorPosition(
+          newNodeModel,
+          edge.endPoint,
+          newNodeModel.width,
+          newNodeModel.height,
+        );
+        edge.updateEndPoint(point);
+      }
+    });
+  }
+
   /* 设置主题 */
   @action setTheme(style: Style) {
     this.theme = updateTheme({ ...this.theme, ...style });